# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicianProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile with registration status",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile": {
      "patch": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Profile fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicianProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update musician profile (name, instrument, level, contact, phone, isHost)",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/inscricoes": {
      "post": {
        "operationId": "InscricaoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration created successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new registration",
        "tags": [
          "Registrations"
        ]
      }
    },
    "/inscricoes/{id}": {
      "patch": {
        "operationId": "InscricaoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - host only"
          },
          "404": {
            "description": "Registration not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update registration details (instrument, status)",
        "tags": [
          "Registrations"
        ]
      },
      "delete": {
        "operationId": "InscricaoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration cancelled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - can only delete own registrations"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel registration",
        "tags": [
          "Registrations"
        ]
      }
    },
...
    "/feedback": {
      "get": {
        "operationId": "FeedbackController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated feedback list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - host role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all feedback (host only)",
        "tags": [
          "Feedback"
        ]
      },
      "post": {
        "operationId": "FeedbackController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Submit feedback (authentication optional)",
        "tags": [
          "Feedback"
        ]
      }
    },
...
  },
  "info": {
    "title": "Karaoke Jam API",
    "description": "Backend API para organização de Jam Sessions presenciais em tempo real",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MusicianProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Musician ID (UUID)"
          },
          "supabaseUserId": {
            "type": "string",
            "description": "Supabase user ID"
          },
          "name": {
            "type": "string",
            "description": "Musician name"
          },
          "email": {
            "type": "string",
            "description": "Musician email"
          },
          "phone": {
            "type": "string",
            "description": "Musician phone"
          },
          "contact": {
            "type": "string",
            "description": "Primary contact info"
          },
          "instrument": {
            "type": "string",
            "description": "Instrument (drums, guitar, vocals, bass, keys, etc.)"
          },
          "level": {
            "type": "string",
            "description": "Skill level",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "PROFESSIONAL"
            ]
          },
          "isHost": {
            "type": "boolean",
            "description": "Whether user is a jam host"
          },
          "registrationComplete": {
            "type": "boolean",
            "description": "Whether profile registration is complete (has instrument, level, and contact)"
          },
          "isNewUser": {
            "type": "boolean",
            "description": "Whether is new user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp"
          }
        },
        "required": [
          "id",
          "supabaseUserId",
          "name",
          "email",
          "phone",
          "contact",
          "instrument",
          "level",
          "isHost",
          "registrationComplete",
          "isNewUser",
          "createdAt"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Musician name"
          },
          "instrument": {
            "type": "string",
            "description": "Musician instrument (drums, guitar, vocals, bass, keys, etc.)"
          },
          "level": {
            "type": "string",
            "description": "Musician skill level",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED",
              "PROFESSIONAL"
            ]
          },
          "contact": {
            "type": "string",
            "description": "Contact phone number or secondary email"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          }
        }
      },
...
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "string",
            "description": "Schedule ID"
          },
          "musicianId": {
            "type": "string",
            "description": "Musician ID (optional, only usable by hosts)"
          },
          "instrument": {
            "type": "string",
            "description": "Instrument to play"
          }
        },
        "required": [
          "scheduleId"
        ]
      },
      "UpdateRegistrationDto": {
        "type": "object",
        "properties": {
          "instrument": {
            "type": "string",
            "description": "Instrument to play"
          },
          "status": {
            "type": "string",
            "description": "Registration status",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          }
        }
      },
...
      "FeedbackListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Feedback ID (UUID)"
          },
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5"
          },
          "comment": {
            "type": "string",
            "description": "Optional comment"
          },
          "pageUrl": {
            "type": "string",
            "description": "Page URL where feedback was submitted"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "musicianId": {
            "type": "string",
            "description": "Musician ID if authenticated"
          },
          "musicianName": {
            "type": "string",
            "description": "Musician name if authenticated"
          }
        },
        "required": [
          "id",
          "rating",
          "createdAt"
        ]
      },
      "FeedbackListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of feedback items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedbackListItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of feedback items"
          },
          "page": {
            "type": "number",
            "description": "Current page number"
          },
          "limit": {
            "type": "number",
            "description": "Items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Optional comment",
            "maxLength": 500
          },
          "userAgent": {
            "type": "string",
            "description": "Browser user agent"
          },
          "pageUrl": {
            "type": "string",
            "description": "Page URL where feedback was submitted"
          }
        },
        "required": [
          "rating"
        ]
      },
      "FeedbackResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Feedback ID (UUID)"
          },
          "rating": {
            "type": "number",
            "description": "Rating from 1 to 5"
          },
          "comment": {
            "type": "string",
            "description": "Optional comment"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "musicianId": {
            "type": "string",
            "description": "Musician ID if authenticated"
          }
        },
        "required": [
          "id",
          "rating",
          "createdAt"
        ]
      },
...
    }
  }
}